SW-2913 Move country search api into location service
- rename i18service to location service which handles countries and time zones
