+++ src/providers/LocalizationProvider.tsx
@@ -4,7 +4,7 @@ import { TimeZoneDescription } from 'src/types/TimeZones';
 import strings, { ILocalizedStringsMap } from 'src/strings';
 import { ProvidedLocalizationData, useUser } from '.';
 import { supportedLocales } from '../strings/locales';
-import { HttpService, I18nService } from 'src/services';
+import { HttpService, LocationService } from 'src/services';
 
 export type LocalizationProviderProps = {
   children?: React.ReactNode;
@@ -37,7 +37,7 @@ export default function LocalizationProvider({
   // This must come after the effect that configures the Accept-Language header.
   useEffect(() => {
     const fetchTimeZones = async () => {
-      const timeZoneResponse = await I18nService.getTimeZones();
+      const timeZoneResponse = await LocationService.getTimeZones();
       if (!timeZoneResponse.error && timeZoneResponse.timeZones) {
         setTimeZones(timeZoneResponse.timeZones.sort((a, b) => a.longName.localeCompare(b.longName, selectedLocale)));
       }
